Consolidate and update samples
