+++ src/Xunit.StaFact.Tests/Xunit.StaFact.Tests.ruleset
@@ -75,5 +75,6 @@
     <Rule Id="SA1130" Action="None" />
     <Rule Id="SA1600" Action="None" />
     <Rule Id="SA1601" Action="None" />
+    <Rule Id="SA1615" Action="Hidden" />
   </Rules>
 </RuleSet>
